change code style to mvc: add controllers
