+++ app.js
@@ -1,35 +1,42 @@
 var express = require('express');
 var path = require('path');
 var favicon = require('serve-favicon');
+
 var logger = require('morgan');
+
 var bodyParser = require('body-parser');
 var cookieParser = require('cookie-parser');
-var session = require('express-session');
-//var mongoStore = require('connect-mongo')(session);
+ // var session = require('express-session');
+// var MongoStore = require('connect-mongo')(session);
+
 var ejs = require('ejs');
 var routes = require('./routes/index');
+
 var app = express();
 
-// view engine setup
+app.use(favicon(__dirname + '/public/images/favicon.ico'));
+
 app.set('views', path.join(__dirname, 'views'));
 app.engine('.html', ejs.__express);
 app.set('view engine', 'html');
 
 app.use(logger('dev'));
-app.use(bodyParser.json());
 
 app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+
+app.use(express.static(path.join(__dirname, 'public')));
+
 app.use(cookieParser());
 // app.use(session({
-//   resave:false,
-//   saveUninitialized: true,
-//   secret: 'issues'
-//   // store: new mongoStore({
-//   //   url:process.env.DB_HOST,
-//   //   collection: 'sessions'
-//   // })
+//   secret: 'im',
+  // store: new MongoStore({
+  //   url:"mongodb://localhost/issuesManag",
+  //   collection:"sessions"
+  // }),
+//   resave: false,
+//   saveUninitialized: true
 // }))
-app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', routes);
 
@@ -41,7 +48,6 @@ app.use(function(req, res, next) {
 });
 
 // error handlers
-
 // development error handler
 // will print stacktrace
 if (app.get('env') === 'development') {
