+++ package.json
@@ -9,13 +9,11 @@
     "bluebird": "^3.3.4",
     "body-parser": "~1.13.2",
     "connect": "^3.4.1",
-    "connect-mongo": "^1.1.0",
     "cookie-parser": "~1.3.5",
     "debug": "^2.2.0",
     "ejs": "~2.3.3",
     "env2": "^2.0.7",
     "express": "~4.13.1",
-    "express-session": "^1.13.0",
     "handlebars": "^4.0.5",
     "moment": "^2.13.0",
     "mongoose": "^4.4.11",
